Add movement model

+++ database/migrations/2019_04_15_191332603432_create_1555355612603_permission_role_pivot_table.php
@@ -9,9 +9,9 @@ class Create1555355612603PermissionRolePivotTable extends Migration
     public function up()
     {
         Schema::create('permission_role', function (Blueprint $table) {
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->foreign('role_id')->references('id')->on('roles');
-            $table->unsignedInteger('permission_id');
+            $table->unsignedBigInteger('permission_id');
             $table->foreign('permission_id')->references('id')->on('permissions');
         });
     }
